Reorganize test structure into tools/ and functions/ subdirectories
**Test Structure Changes:**
- Split unit tests into `tools/` (MCP tools) and `functions/` (helper functions)
- Split monolithic `build-structured-query.test.ts` into 6 separate function-specific files
- Organized integration tests under `tools/` subdirectory for consistency

**New Structure:**
```
tests/
├── unit/
│   ├── tools/        # 6 MCP tool tests (list-sources, query-logs, etc.)
│   └── functions/    # 6 helper function tests (buildStructuredQuery, etc.)
└── integration/
    └── tools/        # 7 MCP tool integration tests
```

**Benefits:**
- Better organization and easier navigation
- Maintains one-function-per-file testing standard
- Clear separation between MCP tools and helper functions
- All 73 unit tests continue passing
- Updated all relative import paths for new directory structure

**Files Split:**
- `build-structured-query.test.ts` → 6 dedicated function test files
- `validate-query-params.test.ts` (new)
- `sanitize-sql-string.test.ts` (new)
- `validate-json-field-filter.test.ts` (new)
- `parse-relative-time.test.ts` (new)
- `parse-time-value.test.ts` (new)
