+++ tests/unit/functions/build-structured-query.test.ts
@@ -1,17 +1,7 @@
 import { describe, it, expect } from 'vitest'
+import { buildStructuredQuery, type StructuredQueryParams } from '../../../src/tools/query-tools.js'
 
-// Import the actual structured query functions for direct testing
-import {
-  buildStructuredQuery,
-  validateQueryParams,
-  sanitizeSqlString,
-  validateJsonFieldFilter,
-  parseRelativeTime,
-  parseTimeValue,
-  type StructuredQueryParams
-} from '../../src/tools/query-tools.js'
-
-describe('buildStructuredQuery Function (Simplified Schema)', () => {
+describe('buildStructuredQuery Function', () => {
   describe('Basic Query Generation', () => {
     it('should generate basic query with default parameters', async () => {
       const params: StructuredQueryParams = {
@@ -391,153 +381,49 @@ describe('buildStructuredQuery Function (Simplified Schema)', () => {
       expect(query).toBe("SELECT dt, raw, json, getJSON(raw, 'level') as level FROM logs ORDER BY dt DESC LIMIT 5")
     })
   })
-})
 
-describe('validateQueryParams Function (Simplified)', () => {
-  it('should validate valid parameters', () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'raw'],
-      limit: 10
-    }
+  describe('Error Conditions and Edge Cases', () => {
+    it('should throw error for invalid log level', async () => {
+      const params: StructuredQueryParams = {
+        fields: ['dt', 'raw'],
+        filters: {
+          level: 'INVALID_LEVEL' as any
+        },
+        limit: 10
+      }
 
-    expect(() => validateQueryParams(params)).not.toThrow()
-  })
-
-  it('should reject invalid fields', () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'invalid_field'],
-      limit: 10
-    }
-
-    expect(() => validateQueryParams(params)).toThrow('Invalid fields: invalid_field')
-  })
-
-  it('should reject invalid limit (too low)', () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'raw'],
-      limit: 0
-    }
-
-    expect(() => validateQueryParams(params)).toThrow('Invalid limit: 0. Must be between 1 and 1000')
-  })
-
-  it('should reject invalid limit (too high)', () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'raw'],
-      limit: 1001
-    }
-
-    expect(() => validateQueryParams(params)).toThrow('Invalid limit: 1001. Must be between 1 and 1000')
-  })
-})
-
-describe('Other Helper Functions (Unchanged)', () => {
-  describe('sanitizeSqlString Function', () => {
-    it('should escape single quotes', () => {
-      expect(sanitizeSqlString("test'value")).toBe("test''value")
+      await expect(buildStructuredQuery(params)).rejects.toThrow('Invalid log level: INVALID_LEVEL')
     })
 
-    it('should handle multiple single quotes', () => {
-      expect(sanitizeSqlString("user's 'special' data")).toBe("user''s ''special'' data")
+    it('should handle edge case limits', async () => {
+      // Test minimum limit
+      const params1: StructuredQueryParams = {
+        fields: ['dt'],
+        limit: 1
+      }
+      const query1 = await buildStructuredQuery(params1)
+      expect(query1).toBe('SELECT dt FROM logs ORDER BY dt DESC LIMIT 1')
+
+      // Test maximum limit
+      const params2: StructuredQueryParams = {
+        fields: ['dt'],
+        limit: 1000
+      }
+      const query2 = await buildStructuredQuery(params2)
+      expect(query2).toBe('SELECT dt FROM logs ORDER BY dt DESC LIMIT 1000')
     })
 
-    it('should handle SQL injection attempts', () => {
-      const maliciousInput = "'; DROP TABLE logs; --"
-      const sanitized = sanitizeSqlString(maliciousInput)
-      expect(sanitized).toBe("''; DROP TABLE logs; --")
-      expect(sanitized).not.toMatch(/^[^']*'[^']*;/)
+    it('should handle special characters in filter values', async () => {
+      const params: StructuredQueryParams = {
+        fields: ['dt', 'raw'],
+        filters: {
+          raw_contains: "user's data with \"quotes\" and \nnewlines"
+        },
+        limit: 10
+      }
+
+      const query = await buildStructuredQuery(params)
+      expect(query).toContain("raw LIKE '%user''s data with \"quotes\" and \nnewlines%'")
     })
-  })
-
-  describe('validateJsonFieldFilter Function', () => {
-    it('should validate valid JSON field filter', () => {
-      const jsonField = { path: 'user.id', value: '12345' }
-      expect(() => validateJsonFieldFilter(jsonField)).not.toThrow()
-    })
-
-    it('should reject invalid path format', () => {
-      const jsonField = { path: '.invalid', value: '12345' }
-      expect(() => validateJsonFieldFilter(jsonField)).toThrow('Invalid JSON path format')
-    })
-  })
-
-  describe('parseRelativeTime Function', () => {
-    it('should parse compact format', () => {
-      expect(parseRelativeTime('1h')).toBe('dt >= now() - INTERVAL 1 HOUR')
-      expect(parseRelativeTime('30m')).toBe('dt >= now() - INTERVAL 30 MINUTE')
-      expect(parseRelativeTime('2d')).toBe('dt >= now() - INTERVAL 2 DAY')
-    })
-
-    it('should parse natural language format', () => {
-      expect(parseRelativeTime('1 hour')).toBe('dt >= now() - INTERVAL 1 HOUR')
-      expect(parseRelativeTime('30 minutes')).toBe('dt >= now() - INTERVAL 30 MINUTE')
-    })
-
-    it('should return null for invalid formats', () => {
-      expect(parseRelativeTime('invalid')).toBeNull()
-      expect(parseRelativeTime('1x')).toBeNull()
-    })
-  })
-
-  describe('parseTimeValue Function', () => {
-    it('should parse ISO date format', () => {
-      expect(parseTimeValue('2024-01-15', 'start')).toBe("dt >= '2024-01-15'")
-      expect(parseTimeValue('2024-01-15', 'end')).toBe("dt <= '2024-01-15'")
-    })
-
-    it('should return null for invalid formats', () => {
-      expect(parseTimeValue('invalid-date', 'start')).toBeNull()
-      expect(parseTimeValue('not-a-date', 'end')).toBeNull()
-    })
-
-    it('should accept format-valid dates (semantic validation is database responsibility)', () => {
-      expect(parseTimeValue('2024-13-01', 'start')).toBe("dt >= '2024-13-01'")
-    })
-  })
-})
-
-describe('Error Conditions and Edge Cases', () => {
-  it('should throw error for invalid log level', async () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'raw'],
-      filters: {
-        level: 'INVALID_LEVEL' as any
-      },
-      limit: 10
-    }
-
-    await expect(buildStructuredQuery(params)).rejects.toThrow('Invalid log level: INVALID_LEVEL')
-  })
-
-
-  it('should handle edge case limits', async () => {
-    // Test minimum limit
-    const params1: StructuredQueryParams = {
-      fields: ['dt'],
-      limit: 1
-    }
-    const query1 = await buildStructuredQuery(params1)
-    expect(query1).toBe('SELECT dt FROM logs ORDER BY dt DESC LIMIT 1')
-
-    // Test maximum limit
-    const params2: StructuredQueryParams = {
-      fields: ['dt'],
-      limit: 1000
-    }
-    const query2 = await buildStructuredQuery(params2)
-    expect(query2).toBe('SELECT dt FROM logs ORDER BY dt DESC LIMIT 1000')
-  })
-
-  it('should handle special characters in filter values', async () => {
-    const params: StructuredQueryParams = {
-      fields: ['dt', 'raw'],
-      filters: {
-        raw_contains: "user's data with \"quotes\" and \nnewlines"
-      },
-      limit: 10
-    }
-
-    const query = await buildStructuredQuery(params)
-    expect(query).toContain("raw LIKE '%user''s data with \"quotes\" and \nnewlines%'")
   })
 })
